use getEnumerationValue for GasType

+++ src/EnergyPlus/WindowManager.cc
@@ -7539,7 +7539,7 @@ namespace WindowManager {
                             print(state.files.eio,
                                   Format_702,
                                   thisMaterial->Name,
-                                  GasTypeName(thisMaterial->GasType(1)),
+                                  GasTypeName(static_cast<int>(thisMaterial->GasType(1))),
                                   thisMaterial->Thickness);
                             //! fw CASE(WindowGasMixture)
                         } break;
@@ -7746,7 +7746,7 @@ namespace WindowManager {
                             print(state.files.eio,
                                   Format_713,
                                   thisMaterial->Name,
-                                  GasTypeName(thisMaterial->GasType(1)),
+                                  GasTypeName(static_cast<int>(thisMaterial->GasType(1))),
                                   thisMaterial->Thickness,
                                   GapVentType);
                         } break;
